--- a/lib/constants.js
+++ b/lib/constants.js
@@ -20,8 +20,5 @@ module.exports = {
     STACKABLE_LIMIT: 100,
     SPATIALMATCH_STACK_LIMIT: 30,
     MAX_CORRECTION_LENGTH: 8,
-    VERIFYMATCH_STACK_LIMIT: 20,
-    // Regexp pattern to determine word boundaries when making token replacements
-    // detects all punctuation including unicode punctuation plus all whitespace
-    WORD_BOUNDARY: "[\\s\\u2000-\\u206F\\u2E00-\\u2E7F\\\\'!\"#$%&()*+,\\-.\\/:;<=>?@\\[\\]^_`{|}~]"
+    VERIFYMATCH_STACK_LIMIT: 20
 };