--- a/lib/indexer/indexdocs.js
+++ b/lib/indexer/indexdocs.js
@@ -66,7 +66,7 @@ function indexdocs(docs, source, options, callback) {
     if (TIMER) console.time('update:freq');
     let freq;
     try {
-        freq = generateFrequency(docs, source.simple_replacer.concat(source.complex_query_replacer), options.tokens, parseInt(source.maxscore), source.lang.has_languages);
+        freq = generateFrequency(docs, source.simple_replacer, source.complex_query_replacer, options.tokens, parseInt(source.maxscore), source.lang.has_languages);
     } catch (err) {
         return callback(err);
     }
@@ -431,6 +431,13 @@ function loadDoc(freq, patch, doc, source, zoom, simple_replacer, complex_replac
     return doc;
 }
 
+/**
+ * Detect whether a set of tiles contains a location
+ *
+ * @param {Array<number>} center - 2 element lon,lat array .
+ * @param {Array<Array<number>>} tiles - array of x,y,z arrays
+ * @return {boolean}
+ */
 function verifyCenter(center, tiles) {
     let found = false;
     let i = 0;
@@ -444,7 +451,17 @@ function verifyCenter(center, tiles) {
     return found;
 }
 
-function generateFrequency(docs, replacer, globalTokens, maxScore) {
+/**
+ * Generate term frequency
+ *
+ * @param {Array} docs
+ * @param {Object} simpleReplacer
+ * @param {Array} complexReplacer
+ * @param {Array} globalTokens
+ * @param {Number} maxScore
+ * @return {Object}
+ */
+function generateFrequency(docs, simpleReplacer, complexReplacer, globalTokens, maxScore) {
     const freq = {};
     // Total # of docs.
     freq['__COUNT__'] = [0];
@@ -459,7 +476,7 @@ function generateFrequency(docs, replacer, globalTokens, maxScore) {
         // set max score
         freq['__MAX__'][0] = maxScore || Math.max(freq['__MAX__'][0], docs[i].properties['carmen:score'] || 0);
 
-        const texts = termops.getMinimalIndexableText(replacer, globalTokens, docs[i]);
+        const texts = termops.getMinimalIndexableText(simpleReplacer, complexReplacer, globalTokens, docs[i]);
 
         for (let x = 0; x < texts.length; x++) {
             const terms = texts[x];