--- a/lib/text-processing/termops.js
+++ b/lib/text-processing/termops.js
@@ -70,6 +70,29 @@ function feature(id) {
     return Math.abs(parseInt(id,10)) % mp20;
 }
 
+const WORD_SEPARATOR = [
+    // Equivalient to \u0020\f\n\r\t\v\u00A0\u1680\u2000-\u200A\u2028\u2029\u202F\u205F\u3000\uFEFF
+    '\\s',
+
+    // \u2000 - \u206F "General Punctuation"
+    '\u200B-\u206F',
+
+    // \u2E00 - \u2E7F "Supplemental Punctuation"
+    '\u2E00-\u2E7F',
+
+    // The usual suspects from \u0020 - \u007F "Basic Latin" except -.
+    // !"#$%&'()*+,
+    '\u0021-\u002C',
+    '\/',
+    // :;<=>?@
+    '\u003A-\u0040',
+    // [\]^_`
+    '\u005B-\u0060',
+    // {|}~
+    '\u007B-\u007E',
+].join('');
+module.exports.WORD_SEPARATOR = WORD_SEPARATOR;
+
 /**
  * tokenize - Normalize input text into lowercase, asciified tokens.
  *
@@ -88,13 +111,10 @@ function tokenize(query, lonlat) {
     }
     const normalized = query
         .toLowerCase()
-        .replace(/[\^]+/g, '')
-        // collapse apostraphes, periods
-        .replace(/[\u2018\u2019\u02BC\u02BB\uFF07'\.]/g, '')
-        // all other ascii and unicode punctuation except '-' per
-        // http://stackoverflow.com/questions/4328500 split terms
-        .replace(/[\u2000-\u206F\u2E00-\u2E7F\\'!"#\$%&\(\)\*\+,\.\/:;<=>\?@\[\]\^_`\{\|\}~]/gi, ' ')
-        .split(/[\s+]+/gi);
+        // collapse apostraphes, periods, caret
+        .replace(/[\u2018\u2019\u02BC\u02BB\uFF07'\.\^]/g, '')
+        // split based on other ascii and unicode punctuation except '-'
+        .split(new RegExp(`[${WORD_SEPARATOR}+]`, 'g'));
 
     let pretokens = [];