--- a/lib/text-processing/termops.js
+++ b/lib/text-processing/termops.js
@@ -113,14 +113,20 @@ function tokenize(query, lonlat) {
         .toLowerCase()
         // collapse apostraphes, periods, caret
         .replace(/[\u2018\u2019\u02BC\u02BB\uFF07'\.\^]/g, '')
-        // split based on other ascii and unicode punctuation except '-'
+        // split based on other ascii and unicode punctuation except '-' which
+        // we handle separately below.
         .split(new RegExp(`[${WORD_SEPARATOR}+]`, 'g'));
 
     let pretokens = [];
 
     for (let k = 0; k < normalized.length; k++) {
-        // keep multi-digit numbers from being split in CJK queries
-        if (normalized[k].length && normalized[k][0].charCodeAt() >= 19968 && normalized[k][0].charCodeAt() <= 40959) {
+        if (normalized[k].length === 0) continue;
+
+        const charCode = normalized[k][0].charCodeAt();
+        if (charCode >= 0x4E00 && charCode <= 0x9FFF) {
+            // Numbers in a word that starts with a \u4E00 - \u9FFF "CJK Unified
+            // Ideographs" character are split into their own token here so they
+            // are proteced from the naieve CJK character by character splitting.
             pretokens = pretokens.concat(normalized[k].split(/([0-9０-９]+)/));
         } else if (/(\d+)-(\d+)[a-z]?/.test(normalized[k])) {
             pretokens.push(normalized[k]);
@@ -138,7 +144,10 @@ function tokenize(query, lonlat) {
         // be rejoined into phrases or subqueries and unidecoded.
         if (pretokens[i].length && removeEmoji(pretokens[i]).length) {
             const charCode = pretokens[i][0].charCodeAt();
-            if (charCode >= 19968 && charCode <= 40959) {
+            // \u4E00 - \u9FFF "CJK Unified Ideographs" characters are indexed
+            // individually to support addresses being written from largest to
+            // smallest geographical entity without delimiters.
+            if (charCode >= 0x4E00 && charCode <= 0x9FFF) {
                 tokens = tokens.concat(pretokens[i].split(''));
             } else {
                 tokens.push(pretokens[i]);