--- a/lib/text-processing/termops.js
+++ b/lib/text-processing/termops.js
@@ -117,44 +117,33 @@ function tokenize(query, lonlat) {
         // we handle separately below.
         .split(new RegExp(`[${WORD_SEPARATOR}+]`, 'g'));
 
-    let pretokens = [];
+    const tokens = [];
 
     for (let k = 0; k < normalized.length; k++) {
         if (normalized[k].length === 0) continue;
+        if (removeEmoji(normalized[k]).length === 0) continue;
 
         const charCode = normalized[k][0].charCodeAt();
         if (charCode >= 0x4E00 && charCode <= 0x9FFF) {
-            // Numbers in a word that starts with a \u4E00 - \u9FFF "CJK Unified
-            // Ideographs" character are split into their own token here so they
-            // are proteced from the naieve CJK character by character splitting.
-            pretokens = pretokens.concat(normalized[k].split(/([0-9０-９]+)/));
-        } else if (/(\d+)-(\d+)[a-z]?/.test(normalized[k])) {
-            pretokens.push(normalized[k]);
-        } else {
-            const splitPhrase = normalized[k].split('-');
-            pretokens = pretokens.concat(splitPhrase);
-        }
-    }
-
-    let tokens = [];
-
-    for (let i = 0; i < pretokens.length; i++) {
-        // Exclude any empty strings or strings that will be empty after
-        // emoji-stripping. This is applied as tokenized strings will eventually
-        // be rejoined into phrases or subqueries and unidecoded.
-        if (pretokens[i].length && removeEmoji(pretokens[i]).length) {
-            const charCode = pretokens[i][0].charCodeAt();
             // \u4E00 - \u9FFF "CJK Unified Ideographs" characters are indexed
             // individually to support addresses being written from largest to
-            // smallest geographical entity without delimiters.
-            if (charCode >= 0x4E00 && charCode <= 0x9FFF) {
-                tokens = tokens.concat(pretokens[i].split(''));
-            } else {
-                tokens.push(pretokens[i]);
+            // smallest geographical entity without delimiters. Adjacent numbers,
+            // normal and full-width, are not split.
+            const subtoken = normalized[k].split(/([0-9０-９]+)/);
+            for (let l = 0; l < subtoken.length; l++) {
+                const c = subtoken[l].charCodeAt();
+                if (c >= 0x4E00 && c <= 0x9FFF) {
+                    Array.prototype.push.apply(tokens, [...subtoken[l]]);
+                } else if (subtoken[l].length > 0) {
+                    tokens.push(subtoken[l]);
+                }
             }
+        } else if (/(\d+)-(\d+)[a-z]?/.test(normalized[k])) {
+            tokens.push(normalized[k]);
+        } else {
+            Array.prototype.push.apply(tokens, normalized[k].split('-').filter((v) => v.length > 0));
         }
     }
-
     return tokens;
 }