--- a/lib/text-processing/termops.js
+++ b/lib/text-processing/termops.js
@@ -97,7 +97,7 @@ module.exports.WORD_SEPARATOR = WORD_SEPARATOR;
  * tokenize - Normalize input text into lowercase, asciified tokens.
  *
  * @param  {String} query  A string to tokenize
- * @param  {String} lonlat A lon,lat pair to tokenize
+ * @param  {Boolean} lonlat Whether to attempt lon,lat parse
  * @return {Array}         A tokenized array
  */
 function tokenize(query, lonlat) {
@@ -367,7 +367,7 @@ function getIndexableText(simpleReplacer, complexReplacer, globalReplacer, doc, 
 // Takes a geocoder_tokens token mapping and a text string and returns
 // an array of one or more arrays of tokens that should be indexed.
 module.exports.getMinimalIndexableText = getMinimalIndexableText;
-function getMinimalIndexableText(replacer, globalTokens, doc) {
+function getMinimalIndexableText(simpleReplacer, complexReplacer, globalReplacer, doc) {
     const uniqTexts = new Set();
     const indexableText = [];
     const housenumRange = getHousenumRangeV3(doc);
@@ -380,11 +380,18 @@ function getMinimalIndexableText(replacer, globalTokens, doc) {
         }
     }
 
+    let combinedReplacer = [];
+    if (globalReplacer && globalReplacer.length) combinedReplacer = combinedReplacer.concat(globalReplacer);
+    if (complexReplacer && complexReplacer.length) combinedReplacer = combinedReplacer.concat(complexReplacer);
+
     for (let x = 0; x < texts.length; x++) {
-        if (globalTokens) {
-            texts[x] = token.replaceToken(globalTokens, texts[x]).query;
+        texts[x] = token.replaceToken(combinedReplacer, texts[x]).query;
+
+        // do simple token replacements without regexes -- just look them up in a dict
+        let tokens = tokenize(texts[x]);
+        if (simpleReplacer && simpleReplacer.tokens) {
+            tokens = tokens.map((token) => simpleReplacer.tokens.get(token) || token);
         }
-        const tokens = tokenize(token.replaceToken(replacer, texts[x]).query);
         if (!tokens.length) continue;
 
         if (housenumRange) {