--- a/lib/text-processing/termops.js
+++ b/lib/text-processing/termops.js
@@ -70,12 +70,15 @@ function feature(id) {
     return Math.abs(parseInt(id,10)) % mp20;
 }
 
-const WORD_SEPARATOR = [
+// Split queries based on other ascii and unicode punctuation except '-' which
+// we handle separately.
+const WORD_SEPARATOR = new RegExp([
+    '[',
     // Equivalient to \u0020\f\n\r\t\v\u00A0\u1680\u2000-\u200A\u2028\u2029\u202F\u205F\u3000\uFEFF
     '\\s',
 
     // \u2000 - \u206F "General Punctuation"
-    '\u200B-\u206F',
+    '\u2000-\u206F',
 
     // \u2E00 - \u2E7F "Supplemental Punctuation"
     '\u2E00-\u2E7F',
@@ -90,7 +93,9 @@ const WORD_SEPARATOR = [
     '\u005B-\u0060',
     // {|}~
     '\u007B-\u007E',
-].join('');
+
+    ']'
+].join(''), 'g');
 module.exports.WORD_SEPARATOR = WORD_SEPARATOR;
 
 /**
@@ -113,9 +118,7 @@ function tokenize(query, lonlat) {
         .toLowerCase()
         // collapse apostraphes, periods, caret
         .replace(/[\u2018\u2019\u02BC\u02BB\uFF07'\.\^]/g, '')
-        // split based on other ascii and unicode punctuation except '-' which
-        // we handle separately below.
-        .split(new RegExp(`[${WORD_SEPARATOR}+]`, 'g'));
+        .split(WORD_SEPARATOR);
 
     const tokens = [];