--- a/lib/text-processing/token.js
+++ b/lib/text-processing/token.js
@@ -1,6 +1,33 @@
 'use strict';
 const removeDiacritics = require('./remove-diacritics');
-const WORD_BOUNDARY = require('../constants.js').WORD_BOUNDARY;
+
+// Regexp pattern to determine word boundaries when making token replacements
+// detects punctuation including and whitespace
+const WORD_BOUNDARY = [
+    '[',
+
+    // Equivalient to \u0020\f\n\r\t\v\u00A0\u1680\u2000-\u200A\u2028\u2029\u202F\u205F\u3000\uFEFF
+    '\\s',
+
+    // \u2000 - \u206F "General Punctuation"
+    '\u2000-\u206F',
+
+    // \u2E00 - \u2E7F "Supplemental Punctuation"
+    '\u2E00-\u2E7F',
+
+    // The usual suspects from \u0020 - \u007F "Basic Latin"
+    // !"#$%&'()*+,-./
+    '\u0021-\u002F',
+    // :;<=>?@
+    '\u003A-\u0040',
+    // [\]^_`
+    '\u005B-\u0060',
+    // {|}~
+    '\u007B-\u007E',
+
+    ']'
+].join('');
+module.exports.WORD_BOUNDARY = WORD_BOUNDARY;
 
 /**
  * An individual pattern-based replacement configuration.