--- a/lib/text-processing/token.js
+++ b/lib/text-processing/token.js
@@ -159,44 +159,21 @@ module.exports.createComplexReplacer = function(tokens, inverseOpts) {
  * @returns {Array<ReplaceRule>} an array of replace rules
  */
 module.exports.createSimpleReplacer = function(tokens) {
-    tokens = JSON.parse(JSON.stringify(tokens));
-    const _tokens = new Map();
+    const replacements = new Map();
     if (Array.isArray(tokens)) {
         for (const t of tokens) {
-            _tokens.set(t.from.toLowerCase(), t.to.toLowerCase());
+            replacements.set(t.from.toLowerCase(), t.to.toLowerCase());
         }
-        tokens = _tokens;
     } else {
         Object.keys(tokens).forEach((k) => {
-            _tokens.set(k.toLowerCase(), tokens[k].toLowerCase());
+            replacements.set(k.toLowerCase(), tokens[k].toLowerCase());
         });
-        tokens = _tokens;
     }
 
-    const wb_split = new RegExp(`(${WORD_BOUNDARY})`);
-    const toFunc = function() {
-        return arguments[0]
-            .split(wb_split)
-            .map((segment) => tokens.get(segment.toLowerCase()) || segment)
-            .join('');
+    return {
+        tokens: replacements,
+        replacer: (v) => v.map((word) => replacements.get(word) || word)
     };
-
-    // Token replacer with word boundaries
-    let out;
-    if (tokens.size) {
-        out = [{
-            from: new RegExp(
-                `(${WORD_BOUNDARY}|^)((` +
-                Array.from(tokens.keys()).sort().join('|') +
-                `)(${WORD_BOUNDARY}|$))+`,
-                'gi'),
-            to: toFunc
-        }];
-    } else {
-        out = [];
-    }
-    out.tokens = tokens;
-    return out;
 };
 
 /**
@@ -221,6 +198,10 @@ module.exports.replaceToken = function(tokens, query) {
     return { query: abbr, lastWord: lastWord };
 };
 
+/**
+ * @param {Array} tokens
+ * @param {String} query
+ */
 module.exports.enumerateTokenReplacements = function(tokens, query) {
     // maintain two lists of the output, one ordered (because the order matters)
     // and one set, so we can easily test to see if the new string is one we've