--- a/test/unit/text-processing/termops.getMinimalIndexableText.test.js
+++ b/test/unit/text-processing/termops.getMinimalIndexableText.test.js
@@ -13,33 +13,33 @@ test('termops.getMinimalIndexableText', (t) => {
     texts = [
         ['main', 'street']
     ];
-    t.deepEqual(termops.getMinimalIndexableText(replacer, [], doc), texts, 'creates indexableText');
+    t.deepEqual(termops.getMinimalIndexableText(replacer, [], [], doc), texts, 'creates indexableText');
 
     replacer = token.createSimpleReplacer({ 'Street':'St' });
     doc = { properties: { 'carmen:text': 'Main Street' } };
     texts = [
         ['main', 'st']
     ];
-    t.deepEqual(termops.getMinimalIndexableText(replacer, [], doc), texts, 'creates contracted phrases using geocoder_tokens');
+    t.deepEqual(termops.getMinimalIndexableText(replacer, [], [], doc), texts, 'creates contracted phrases using geocoder_tokens');
 
     replacer = token.createSimpleReplacer({ 'Street':'St' });
     doc = { properties: { 'carmen:text': 'Main Street, main st' } };
     texts = [
         ['main', 'st']
     ];
-    t.deepEqual(termops.getMinimalIndexableText(replacer, [], doc), texts, 'dedupes phrases');
+    t.deepEqual(termops.getMinimalIndexableText(replacer, [], [], doc), texts, 'dedupes phrases');
 
     replacer = token.createSimpleReplacer({ 'Street':'St', 'Lane':'Ln' });
     doc = { properties: { 'carmen:text': 'Main Street Lane' } };
     texts = [
         ['main', 'st', 'ln']
     ];
-    t.deepEqual(termops.getMinimalIndexableText(replacer, [], doc), texts, 'dedupes phrases');
+    t.deepEqual(termops.getMinimalIndexableText(replacer, [], [], doc), texts, 'dedupes phrases');
 
     replacer = token.createComplexReplacer({ 'dix-huitième':'18e' });
     doc = { properties: { 'carmen:text': 'Avenue du dix-huitième régiment' } };
     texts = [['avenue', 'du', '18e', 'régiment']];
-    t.deepEqual(termops.getMinimalIndexableText(replacer, [], doc), texts, 'hypenated replacement');
+    t.deepEqual(termops.getMinimalIndexableText({}, replacer, [], doc), texts, 'hypenated replacement');
 
     replacer = token.createSimpleReplacer({});
     doc = {
@@ -54,7 +54,7 @@ test('termops.getMinimalIndexableText', (t) => {
         ['##', 'main', 'street'],
         ['#', 'main', 'street'],
     ];
-    t.deepEqual(termops.getMinimalIndexableText(replacer, [],  doc), texts, 'with range');
+    t.deepEqual(termops.getMinimalIndexableText(replacer, [], [],  doc), texts, 'with range');
 
     // sets indexDegens to false for translated text
     replacer = token.createSimpleReplacer({});
@@ -63,7 +63,7 @@ test('termops.getMinimalIndexableText', (t) => {
         ['main', 'street'],
         ['el', 'main', 'street']
     ];
-    t.deepEqual(termops.getMinimalIndexableText(replacer, [], doc), texts, 'creates indexableText');
+    t.deepEqual(termops.getMinimalIndexableText(replacer, [], [], doc), texts, 'creates indexableText');
 
     // doesn't indexDegens for synonyms
     replacer = token.createSimpleReplacer({});
@@ -72,7 +72,7 @@ test('termops.getMinimalIndexableText', (t) => {
         ['latveria'],
         ['republic', 'of', 'latveria']
     ];
-    t.deepEqual(termops.getMinimalIndexableText(replacer, [], doc), texts, 'creates indexableText w/ synonyms');
+    t.deepEqual(termops.getMinimalIndexableText(replacer, [], [], doc), texts, 'creates indexableText w/ synonyms');
 
     t.end();
 });