--- a/test/unit/text-processing/termops.tokenize.test.js
+++ b/test/unit/text-processing/termops.tokenize.test.js
@@ -19,16 +19,26 @@ test('tokenizes basic strings', (t) => {
     t.deepEqual(termops.tokenize('foo)bar'), ['foo', 'bar'], 'splits on )');
     t.deepEqual(termops.tokenize('foo b.a.r'), ['foo', 'bar'], 'collapses .');
     t.deepEqual(termops.tokenize('foo\'s bar'), ['foos', 'bar'], 'collapses apostraphe');
-    t.deepEqual(termops.tokenize('69-150'), ['69-150']);
-    t.deepEqual(termops.tokenize('4-10'), ['4-10']);
-    t.deepEqual(termops.tokenize('5-02A'), ['5-02a']);
-    t.deepEqual(termops.tokenize('23-'), ['23']);
-    t.deepEqual(termops.tokenize('San José'), ['san', 'josé']);
-    t.deepEqual(termops.tokenize('Chamonix-Mont-Blanc'), ['chamonix','mont','blanc']);
-    t.deepEqual(termops.tokenize('Москва'), ['москва']);
-    t.deepEqual(termops.tokenize('京都市'), ['京','都','市']);
+    t.deepEqual(termops.tokenize('69-150'), ['69-150'], 'preserves - (numeric)');
+    t.deepEqual(termops.tokenize('4-10'), ['4-10'], 'preserves - (numeric, short)');
+    t.deepEqual(termops.tokenize('5-02A'), ['5-02a'], 'preserves - (numeric w/ alpha suffix)');
+    t.deepEqual(termops.tokenize('23-'), ['23'], 'strips trailing -');
+    t.deepEqual(termops.tokenize('San José'), ['san', 'josé'], 'preserves accented e');
+    t.deepEqual(termops.tokenize('Chamonix-Mont-Blanc'), ['chamonix','mont','blanc'], 'splits on -');
+    t.deepEqual(termops.tokenize('123, route de N^'), ['123','route','de', 'n'], 'removes ^');
+    t.deepEqual(termops.tokenize('123, route de Nîmes'), ['123','route','de', 'nîmes'], 'preserves î');
+    t.deepEqual(termops.tokenize('Unit 21/2-4'), ['unit','21','2-4']);
+    t.deepEqual(termops.tokenize('Москва'), ['москва'], 'Cyrillic, converts to lower case');
+    t.deepEqual(termops.tokenize('Москва Русский'), ['москва', 'русский'], 'Cyrillic, splits words');
+    t.deepEqual(termops.tokenize('京都市'), ['京','都','市'], 'Splits CJK');
     t.end();
 });
+test.skip('known failures', (t) => {
+    t.deepEqual(termops.tokenize('7/11+Gwynne+Street'), ['7/11','gwynne','street'], 'preserves / (numeric)');
+    t.deepEqual(termops.tokenize('12/3a+Gordon+close'), ['12/1a','gordon','close'], 'preserves / (numeric w/ alpha)');
+    t.deepEqual(termops.tokenize('34+1/2+s+vermont+avenue+#1'), ['34','1/2','s', 'vermont', 'avenue', '#1']);
+});
+
 test('tokenizes lonlat', (t) => {
     t.deepEqual(termops.tokenize('40,0', true), [40,0]);
     t.deepEqual(termops.tokenize('40.00000,-40.31200', true), [40,-40.312]);
@@ -52,10 +62,15 @@ test('edge cases - empty string', (t) => {
 
 test('tokenize Japanese strings with numeric component', (t) => {
     t.deepEqual(termops.tokenize('中津川市馬籠4571-1'), ['中','津','川','市','馬','籠','4571','-','1'], 'dashed number at end');
+    t.deepEqual(termops.tokenize('中津川市馬籠\uFF14\uFF15\uFF17\uFF11-\uFF11'), ['中','津','川','市','馬','籠','４５７１','-','１'], 'dashed full-width number at end');
     t.deepEqual(termops.tokenize('中津川市4571-1馬籠'), ['中','津','川','市','4571','-','1','馬','籠'], 'dashed number in middle');
+    t.deepEqual(termops.tokenize('中津川市\uFF14\uFF15\uFF17\uFF11-\uFF11馬籠'), ['中','津','川','市','４５７１','-','１','馬','籠'], 'dashed full-width number in middle');
     t.deepEqual(termops.tokenize('中津川市4571馬籠'), ['中','津','川','市','4571','馬','籠'], 'number in middle');
+    t.deepEqual(termops.tokenize('中津川市\uFF14\uFF15\uFF17\uFF11馬籠'), ['中','津','川','市','４５７１','馬','籠'], 'full-width number in middle');
     t.deepEqual(termops.tokenize('中津川市4571馬籠123'), ['中','津','川','市','4571','馬','籠','123'], 'numbers in middle and at end');
+    t.deepEqual(termops.tokenize('中津川市\uFF14\uFF15\uFF17\uFF11馬籠\uFF11\uFF12\uFF13'), ['中','津','川','市','４５７１','馬','籠','１２３'], 'full-width numbers in middle and at end');
     t.deepEqual(termops.tokenize('123中津川市4571馬籠'), ['123中津川市4571馬籠'], 'does not split strings that begin with numbers');
+    t.deepEqual(termops.tokenize('\uFF11\uFF12\uFF13中津川市\uFF14\uFF15\uFF17\uFF11馬籠'), ['１２３中津川市４５７１馬籠'], 'does not split strings that begin with full-width numbers');
     t.end();
 });