… test/unit/text-processing/token.replaceToken.test.js