Goal of this PR is to more explicitly capture when we split user input into tokens and why. We do this in a few spots and they've been pretty dense regular expressions and code point ranges. To improve that this PR does the following things;

1. Calls out the regexp that is used to split a query into token as the `WORD_SEPARATOR` in `termops.js` and documents all the Unicode ranges and characters that it contains.
2. Moves the `WORD_BOUNDARY` definition out of the constants file and into `token.js` which is the only place it is used. It also documents all the  Unicode ranges and characters that it contains.
3. Adds a code comment to the CJK specific token splitting and uses hex ranges to make the intent more clear. 
4. ...and I couldn't help it, I consolidated the two post-splitting token iterations. Tests pass and perf looks better, but it could use close review.
5. Can't stop, Also removes the treatment of simple replacements via regexp. It was only used for frequency calculation and category normalization and was overly complex for that use.

Perf stats from my laptop, posting b/c CJK seems much faster on this branch, not quite sure what to make of that.

On Master:
```
# tokenize
tokenize x 360,489 ops/sec ±0.57% (97 runs sampled)
tokenize - cjk x 372,653 ops/sec ±0.73% (92 runs sampled)
```

This branch
```
# tokenize
tokenize x 389,059 ops/sec ±0.72% (95 runs sampled)
tokenize - cjk x 428,671 ops/sec ±0.83% (93 runs sampled)
```

Refs #815 

cc @mapbox/search